crudActividadAdmin

+++ src/controllers/actividades.controller.js
@@ -16,6 +16,12 @@ const getActividadControllerById = (req, res) => {
 };
 
 const postActividadController = async (req, res) => {
+    
+    const role = req.user.role;
+    if (role !== "admin") {
+        return res.status(403).json({ message: "Acceso denegado. Solo los administradores pueden crear actividades." });
+    }
+
     const { nombre, categoria, descripcion, fecha, horario} = req.body;
     createActividad(nombre, categoria, descripcion, fecha, horario);
     res.status(201).json({
@@ -25,6 +31,12 @@ const postActividadController = async (req, res) => {
 
 const putActividadController = (req, res) => {
     const id = req.params.id;
+    const role = req.user.role;
+    if (role !== "admin") {
+        return res.status(403).json({ message: "Acceso denegado. Solo los administradores pueden crear actividades." });
+    }
+
+
     const updated = updateActividad(id, req.body)
     if (updated) {
         res.status(200).json(updated)
@@ -36,6 +48,12 @@ const putActividadController = (req, res) => {
 }
 
 const deleteActividadController = (req, res) => {
+
+    const role = req.user.role;
+    if (role !== "admin") {
+        return res.status(403).json({ message: "Acceso denegado. Solo los administradores pueden crear actividades." });
+    }
+    
     const deleted = deleteActividad(req.params.id)
     if (deleted) {
         res.status(204).json({
